Use a worker pool to manage resources
Instead of simply forking and joining as many lightweight processes as
there are repositories to process, use a pool of workers to control how
much work is done concurrently. Still spawns many lightweight processes,
but they are contending for a worker from the pool to do the actual
operation. By default, use a pool of 8 works, with 0 overflow.

rebar3 ct passes